requirements.txt
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asgiref==3.9.1
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
blis==1.2.0
boto3==1.37.3
botocore==1.37.3
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.3.1
contractions==0.1.73
coverage==7.8.0
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.2
funcy==2.0
gensim==4.3.3
gevent==24.11.1
geventhttpclient==2.3.3
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
limits==5.2.0
locust==2.34.1
locust-plugins==4.6.0
logtail-python==0.3.3
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moto==5.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
murmurhash==1.0.12
narwhals==1.34.1
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-dbapi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-psycopg2==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
preshed==3.0.9
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyahocorasick==2.1.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyLDAvis==3.4.1
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
responses==0.25.7
rich==13.9.4
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.0.2
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.46.1
sympy==1.13.1
textblob==0.19.0
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.20.0
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
